FIX form declarations
